gadget-io: Make `#![no_std]` compatible
