+++ Cargo.toml
@@ -87,7 +87,7 @@ sc-consensus-aura = { version = "0.35.0", default-features = false }
 sc-consensus-grandpa = { version = "0.20.0", default-features = false }
 sc-executor = { version = "0.33.0", default-features = false }
 sc-service = { version = "0.36.0", default-features = false }
-sc-keystore = { version = "26.0.0", default-features = false }
+sc-keystore = { version = "33.0.0", default-features = false }
 sc-telemetry = { version = "16.0.0", default-features = false }
 sc-transaction-pool-api = { version = "29.0.0", default-features = false }
 sc-cli = { version = "0.37.0", default-features = false }
@@ -151,7 +151,8 @@ failure = "0.1.8"
 futures-util = "0.3.30"
 serde = { version = "1.0.197", default-features = false }
 serde_json = "1.0"
-url = { version = "2.5", default-features = false }
+# TODO: Stop using this fork once https://github.com/servo/rust-url/pull/831 lands
+url = { git = "https://github.com/domenukk/rust-url", branch = "no_std", default-features = false }
 async-trait = "0.1.73"
 async-recursion2 = "1.0.5"
 log = "0.4.20"
@@ -170,7 +171,7 @@ serde_bytes = { version = "0.11.12", default-features = false }
 itertools = "0.12.0"
 strum = "0.26"
 clap = "4.5.13"
-color-eyre = { version = "0.6", features = ["tracing-error", "color-spantrace", "issue-url"] }
+color-eyre = { version = "0.6", features = ["tracing-error", "color-spantrace"] }
 hex-literal = "0.4.1"
 rand = "0.8.5"
 rand_core = { version = "0.6", default-features = false }
@@ -189,7 +190,8 @@ sysinfo = "0.30.12"
 anyhow = "1.0.79"
 libsecp256k1 = "0.7.1"
 rayon = { version = "1.8.0" }
-thiserror = { version = "1" }
+# TODO: Stop using this fork once https://github.com/dtolnay/thiserror/pull/304 lands
+thiserror = { git = "https://github.com/quartiq/thiserror", branch = "no-std", default-features = false }
 elliptic-curve = { version = "0.13.8" }
 aes-gcm = { version = "0.10", default-features = false, features = ["aes", "alloc", "getrandom", "zeroize"] }
 zeroize = { version = "1", default-features = false, features = ["alloc", "zeroize_derive"] }
@@ -223,6 +225,8 @@ hyper = { version = "1.3.1", default-features = false }
 pin-project-lite = "0.2.14"
 http-body-util = "0.1.1"
 backoff = { version = "0.4.0", features = ["tokio"] }
+multiaddr = { version = "0.18.1", default-features = false }
+cfg-if = "1.0.0"
 
 anvil = { git = "https://github.com/foundry-rs/foundry.git" }
 foundry-common = { git = "https://github.com/foundry-rs/foundry.git" }
