+++ gadget-io/Cargo.toml
@@ -7,27 +7,31 @@ description = "Tangle's gadget IO library for writing Tangle blueprints"
 repository.workspace = true
 homepage.workspace = true
 
-[features]
-default = ["std"]
-std = []
-wasm = []
-
 [dependencies]
-tracing = { workspace = true, features = ["log", "attributes"] }
+cfg-if = { workspace = true }
+tracing = { workspace = true, default-features = false, features = ["log", "attributes"] }
 hex = { workspace = true }
 scale-info = { workspace = true}
 parity-scale-codec = { workspace = true }
 serde = { workspace = true }
+sp-application-crypto = { workspace = true, default-features = false, features = ["full_crypto"] }
+sp-core = { workspace = true, features = ["serde", "full_crypto"] }
+multiaddr = { workspace = true, default-features = false }
+url = { workspace = true, default-features = false, features = ["alloc", "serde"] }
+thiserror = { workspace = true }
+tsify = { version = "0.4.5", optional = true }
+wasm-bindgen = { workspace = true, optional = true }
+tokio = { workspace = true, optional = true, features = [
+    "time",
+    "rt",
+    "sync",
+] }
 
 [target.'cfg(target_family = "wasm")'.dependencies]
-getrandom = { workspace = true, features = ["js"] }
-wasm-bindgen = { version = "0.2.92" }
 wasm-bindgen-futures = { version = "0.4.42" }
-tsify = { version = "0.4.5" }
+
 serde-wasm-bindgen = { version = "0.6.5" }
 js-sys = "0.3.69"
-color-eyre = { version = "0.6", features = ["tracing-error", "color-spantrace", "issue-url"] }
-sp-core = { workspace = true, features = ["serde", "full_crypto"] }
 tokio = { workspace = true, features = ["sync", "macros", "io-util", "rt", "time"] }
 wasmtimer = { version = "0.2" }
 
@@ -35,13 +39,22 @@ p256 = { workspace = true, features = ["alloc", "ecdsa"], default-features = fal
 
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
-color-eyre = { version = "0.6", features = ["tracing-error", "color-spantrace", "issue-url"] }
-sp-core = { workspace = true, features = ["default"] }
-sp-application-crypto = { workspace = true }
 sp-keystore = { workspace = true }
 sc-keystore = { workspace = true }
 structopt = "0.3.26"
-tokio = { workspace = true, features = ["full"] }
-url = { workspace = true, features = ["serde"] }
-libp2p = { workspace = true }
-rand = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+    "dep:tokio",
+    "sp-application-crypto/std",
+    "tracing/std",
+    "url/std",
+]
+wasm-bindgen = [
+    "dep:wasm-bindgen",
+]
+typescript = [
+    "dep:tsify",
+    "wasm-bindgen",
+]
