+++ gadget-io/src/imp/standard/keystore.rs
@@ -1,11 +1,9 @@
+use crate::error::{GadgetIoError, Result};
 pub use crate::shared::keystore::SubstrateKeystore;
-use color_eyre;
-use color_eyre::eyre::OptionExt;
 use sp_core::{ecdsa, sr25519, ByteArray};
 use std::path::{Path, PathBuf};
 use tracing;
 
-use color_eyre::Result;
 use sc_keystore::{Keystore, LocalKeystore};
 use sp_keystore::KeystorePtr;
 use std::sync::Arc;
@@ -18,7 +16,8 @@ impl KeystoreContainer {
     pub fn new(config: &KeystoreConfig) -> Result<Self> {
         let keystore = Arc::new(match config {
             KeystoreConfig::Path { path, password } => {
-                LocalKeystore::open(path.clone(), password.clone())?
+                LocalKeystore::open(path.clone(), password.clone())
+                    .map_err(|e| GadgetIoError::Other(e.to_string()))?
             }
             KeystoreConfig::InMemory => LocalKeystore::in_memory(),
         });
@@ -83,25 +82,30 @@ pub mod crypto {
 }
 
 impl SubstrateKeystore for KeystoreConfig {
-    fn ecdsa_key(&self) -> color_eyre::Result<ecdsa::Pair> {
+    // TODO: Return more specific error types
+    fn ecdsa_key(&self) -> Result<ecdsa::Pair> {
         let keystore_container = KeystoreContainer::new(self)?;
         let keystore = keystore_container.local_keystore();
         tracing::debug!("Loaded keystore from path");
         let ecdsa_keys = keystore.ecdsa_public_keys(crypto::role::KEY_TYPE);
 
         if ecdsa_keys.len() != 1 {
-            color_eyre::eyre::bail!(
+            return Err(GadgetIoError::Other(format!(
                 "`role`: Expected exactly one key in ECDSA keystore, found {}",
                 ecdsa_keys.len()
-            );
+            )));
         }
 
-        let role_public_key = crypto::role::Public::from_slice(&ecdsa_keys[0].0)
-            .map_err(|_| color_eyre::eyre::eyre!("Failed to parse public key from keystore"))?;
+        let role_public_key = crypto::role::Public::from_slice(&ecdsa_keys[0].0).map_err(|_| {
+            GadgetIoError::Other(String::from("Failed to parse public key from keystore"))
+        })?;
 
         let role_key = keystore
-            .key_pair::<crypto::role::Pair>(&role_public_key)?
-            .ok_or_eyre("Failed to load key `role` from keystore")?
+            .key_pair::<crypto::role::Pair>(&role_public_key)
+            .map_err(|e| GadgetIoError::Other(e.to_string()))?
+            .ok_or(GadgetIoError::Other(String::from(
+                "Failed to load key `role` from keystore",
+            )))?
             .into_inner();
 
         tracing::debug!(%role_public_key, "Loaded key from keystore");
@@ -111,25 +115,31 @@ impl SubstrateKeystore for KeystoreConfig {
         Ok(role_key)
     }
 
-    fn sr25519_key(&self) -> color_eyre::Result<sr25519::Pair> {
+    // TODO: Return more specific error types
+    fn sr25519_key(&self) -> Result<sr25519::Pair> {
         let keystore_container = KeystoreContainer::new(self)?;
         let keystore = keystore_container.local_keystore();
         tracing::debug!("Loaded keystore from path");
         let sr25519_keys = keystore.sr25519_public_keys(crypto::acco::KEY_TYPE);
 
         if sr25519_keys.len() != 1 {
-            color_eyre::eyre::bail!(
+            return Err(GadgetIoError::Other(format!(
                 "`acco`: Expected exactly one key in SR25519 keystore, found {}",
                 sr25519_keys.len()
-            );
+            )));
         }
 
-        let account_public_key = crypto::acco::Public::from_slice(&sr25519_keys[0].0)
-            .map_err(|_| color_eyre::eyre::eyre!("Failed to parse public key from keystore"))?;
+        let account_public_key =
+            crypto::acco::Public::from_slice(&sr25519_keys[0].0).map_err(|_| {
+                GadgetIoError::Other(String::from("Failed to parse public key from keystore"))
+            })?;
 
         let acco_key = keystore
-            .key_pair::<crypto::acco::Pair>(&account_public_key)?
-            .ok_or_eyre("Failed to load key `acco` from keystore")?
+            .key_pair::<crypto::acco::Pair>(&account_public_key)
+            .map_err(|e| GadgetIoError::Other(e.to_string()))?
+            .ok_or(GadgetIoError::Other(String::from(
+                "Failed to load key `acco` from keystore",
+            )))?
             .into_inner();
 
         tracing::debug!(%account_public_key, "Loaded key from keystore");
