+++ gadget-io/src/shared/keystore.rs
@@ -1,8 +1,8 @@
-use color_eyre;
+use crate::error::Result;
 use sp_core::{ecdsa, sr25519};
 
 pub trait SubstrateKeystore {
-    fn ecdsa_key(&self) -> color_eyre::Result<ecdsa::Pair>;
+    fn ecdsa_key(&self) -> Result<ecdsa::Pair>;
 
-    fn sr25519_key(&self) -> color_eyre::Result<sr25519::Pair>;
+    fn sr25519_key(&self) -> Result<sr25519::Pair>;
 }
