+++ gadget-io/src/shared/mod.rs
@@ -1,1 +1,2 @@
 pub mod keystore;
+pub mod shell;
